requirements.txt
django
djangorestframework
pillow
djangorestframework_simplejwt
requests